packages/kakaswap-contracts/pyproject.toml
[tool.poetry]
name = "kakaswap-contracts"
version = "0.1.0"
description = ""
authors = ["Clément Walter <clement0walter@gmail.com>"]
readme = "README.md"
packages = [{ include = "scripts" }]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.1"
web3 = "^5.31.3"
toml = "^0.10.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
